Add all files
