+++ src/ru/gb/family_tree/family_tree/TreeNode.java
@@ -1,0 +1,21 @@
+package ru.gb.family_tree.family_tree;
+
+import java.io.Serializable;
+import java.time.LocalDate;
+import java.util.List;
+
+public interface TreeNode<T> extends Serializable {
+    void setId(long id);
+    long getId();
+    T getFather();
+    T getMother();
+    boolean addChild(T human);
+    boolean addParent(T human);
+    String getName();
+    LocalDate getDeathDate();
+    LocalDate getBirthDate();
+    List<T> getParents();
+    List<T> getChilder();
+    T getSpouse();
+    void setSpouse(T human);
+}
